Moved follow button into user.dart
