+++ lib/profile.dart
@@ -11,6 +11,8 @@ import 'package:fritter/tweet.dart';
 import 'package:intl/intl.dart';
 import 'package:sqflite/sqflite.dart';
 
+import 'user.dart';
+
 class ProfileScreen extends StatelessWidget {
   final String? id;
   final String username;
@@ -217,79 +219,5 @@ class _ProfileScreenBodyState extends State<ProfileScreenBody> {
   void dispose() {
     super.dispose();
     _sub?.cancel();
-  }
-}
-
-class FollowButton extends StatefulWidget {
-  final String id;
-  final String name;
-  final String screenName;
-  final String? imageUri;
-
-  const FollowButton({Key? key, required this.id, required this.name, required this.screenName, this.imageUri}) : super(key: key);
-
-  @override
-  _FollowButtonState createState() => _FollowButtonState();
-}
-
-class _FollowButtonState extends State<FollowButton> {
-  bool? _followed;
-
-  @override
-  void initState() {
-    super.initState();
-
-    fetchFollowed();
-  }
-
-  Future fetchFollowed() {
-    return isFollowed(int.parse(widget.id))
-        .then((value) => setState(() {
-          this._followed = value;
-        }));
-  }
-
-  Future<bool> isFollowed(int id) async {
-    Database database = await Repository.open();
-
-    var result = await database.rawQuery('SELECT EXISTS (SELECT 1 FROM following WHERE id = ?)', [id]);
-    if (result.isEmpty) {
-      return false;
-    }
-
-    return result.first.values.first == 1
-      ? true
-      : false;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    var id = int.parse(widget.id);
-
-    var followed = _followed;
-    if (followed == null) {
-      return Center(child: CircularProgressIndicator());
-    }
-
-    var icon = followed
-      ? Icon(Icons.person_remove)
-      : Icon(Icons.person_add);
-
-    return IconButton(icon: icon, onPressed: () async {
-      Database database = await Repository.open();
-
-      if (followed) {
-        await database.delete('following', where: 'id = ?', whereArgs: [id]);
-      } else {
-        await database.insert('following', {
-          'id': id,
-          'screen_name': widget.screenName,
-          'name': widget.name,
-          'profile_image_url_https': widget.imageUri
-        });
-      }
-
-      await fetchFollowed();
-    });
   }
 }
