+++ lib/user.dart
@@ -1,6 +1,8 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
+import 'package:sqflite/sqflite.dart';
 
+import 'database.dart';
 import 'profile.dart';
 
 class UserTile extends StatelessWidget {
@@ -44,4 +46,75 @@ class UserTile extends StatelessWidget {
     );
   }
 
+}
+
+class FollowButton extends StatefulWidget {
+  final String id;
+  final String name;
+  final String screenName;
+  final String? imageUri;
+
+  const FollowButton({Key? key, required this.id, required this.name, required this.screenName, this.imageUri}) : super(key: key);
+
+  @override
+  _FollowButtonState createState() => _FollowButtonState();
+}
+
+class _FollowButtonState extends State<FollowButton> {
+  bool? _followed;
+
+  @override
+  void initState() {
+    super.initState();
+
+    fetchFollowed();
+  }
+
+  Future fetchFollowed() {
+    return isFollowed(int.parse(widget.id)).then((value) => setState(() {
+      this._followed = value;
+    }));
+  }
+
+  Future<bool> isFollowed(int id) async {
+    Database database = await Repository.open();
+
+    var result = await database.rawQuery('SELECT EXISTS (SELECT 1 FROM following WHERE id = ?)', [id]);
+    if (result.isEmpty) {
+      return false;
+    }
+
+    return result.first.values.first == 1;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    var id = int.parse(widget.id);
+
+    var followed = _followed;
+    if (followed == null) {
+      return Center(child: CircularProgressIndicator());
+    }
+
+    var icon = followed
+        ? Icon(Icons.person_remove)
+        : Icon(Icons.person_add);
+
+    return IconButton(icon: icon, onPressed: () async {
+      Database database = await Repository.open();
+
+      if (followed) {
+        await database.delete('following', where: 'id = ?', whereArgs: [id]);
+      } else {
+        await database.insert('following', {
+          'id': id,
+          'screen_name': widget.screenName,
+          'name': widget.name,
+          'profile_image_url_https': widget.imageUri
+        });
+      }
+
+      await fetchFollowed();
+    });
+  }
 }
